Move function to .c file
